Split pubnub_hyper transport implementation into multiple files for easier navigation
