+++ pubnub-hyper/src/lib.rs
@@ -67,3 +67,6 @@ pub type PubNub = CorePubNub<DefaultTransport, DefaultRuntime>;
 
 pub mod runtime;
 pub mod transport;
+
+#[macro_use]
+mod macros;
